202602_YouTube_Language_Tutor/requirements.txt
google-genai>=1.0.0
youtube-transcript-api>=0.6.2
